git subrepo pull lib/cpr
subrepo:
  subdir:   "lib/cpr"
  merged:   "4946775"
upstream:
  origin:   "https://github.com/EmreTech/cpr.git"
  branch:   "master"
  commit:   "4946775"
git-subrepo:
  version:  "0.4.3"
  origin:   "https://github.com/ingydotnet/git-subrepo"
  commit:   "2f68596"

+++ lib/cpr/include/cpr/callback.h
@@ -44,13 +44,13 @@ class ProgressCallback {
   public:
     ProgressCallback() = default;
     // NOLINTNEXTLINE(google-explicit-constructor, hicpp-explicit-conversions)
-    ProgressCallback(std::function<bool(size_t downloadTotal, size_t downloadNow,
-                                        size_t uploadTotal, size_t uploadNow)>
+    ProgressCallback(std::function<bool(double downloadTotal, double downloadNow,
+                                        double uploadTotal, double uploadNow)>
                              callback)
             : callback(std::move(callback)) {}
 
-    std::function<bool(size_t downloadTotal, size_t downloadNow, size_t uploadTotal,
-                       size_t uploadNow)>
+    std::function<bool(double downloadTotal, double downloadNow, double uploadTotal,
+                       double uploadNow)>
             callback;
 };
